feat: add RecentConferences component and integrate into ClassroomList
