+++ teacher-webapp/src/pages/ClassroomList.js
@@ -21,7 +21,6 @@ import {
   People as PeopleIcon,
   School as SchoolIcon,
   Logout as LogoutIcon,
-  History as HistoryIcon,
   PlayArrow as PlayArrowIcon,
 } from "@mui/icons-material";
 import { useNavigate } from "react-router-dom";
@@ -32,6 +31,7 @@ import { LoadingSpinner } from "../components/common/LoadingSpinner";
 import { PageContainer } from "../components/layout/PageContainer";
 import { ROUTES } from "../constants/routes";
 import { getSessionHistory } from "../services/sessionHistoryService";
+import RecentConferences from "../components/RecentConferences";
 const ClassroomList = () => {
   const navigate = useNavigate();
   const { logout } = useAuth();
@@ -98,27 +98,6 @@ const ClassroomList = () => {
 
   const handleLogout = () => {
     logout();
-  };
-
-  const formatTimestamp = (timestamp) => {
-    const date = new Date(timestamp);
-    const now = new Date();
-    const diffMs = now - date;
-    const diffMins = Math.floor(diffMs / 60000);
-    const diffHours = Math.floor(diffMs / 3600000);
-    const diffDays = Math.floor(diffMs / 86400000);
-
-    if (diffMins < 1) return "Just now";
-    if (diffMins < 60) return `${diffMins} minute${diffMins > 1 ? "s" : ""} ago`;
-    if (diffHours < 24) return `${diffHours} hour${diffHours > 1 ? "s" : ""} ago`;
-    if (diffDays < 7) return `${diffDays} day${diffDays > 1 ? "s" : ""} ago`;
-    
-    // Format as date for older items
-    return date.toLocaleDateString(undefined, {
-      month: "short",
-      day: "numeric",
-      year: date.getFullYear() !== now.getFullYear() ? "numeric" : undefined,
-    });
   };
 
   const handleSessionClick = (sessionItem) => {
@@ -192,73 +171,6 @@ const ClassroomList = () => {
         </Alert>
       )}
 
-      {/* Recent Conference Sessions Section */}
-      {recentSessions.length > 0 && (
-        <Card sx={{ mb: 3 }}>
-          <CardContent>
-            <Box sx={{ display: "flex", alignItems: "center", gap: 1, mb: 2 }}>
-              <HistoryIcon color="primary" />
-              <Typography variant="h6" fontWeight={600}>
-                Recent Conferences
-              </Typography>
-            </Box>
-            <Box sx={{ display: "flex", flexDirection: "column", gap: 1.5 }}>
-              {recentSessions.map((sessionItem) => (
-                <Box
-                  key={`${sessionItem.groupId}-${sessionItem.timestamp}`}
-                  sx={{
-                    display: "flex",
-                    alignItems: "center",
-                    justifyContent: "space-between",
-                    p: 1.5,
-                    border: "1px solid",
-                    borderColor: "divider",
-                    borderRadius: 1,
-                    "&:hover": {
-                      bgcolor: "action.hover",
-                      cursor: "pointer",
-                    },
-                  }}
-                  onClick={() => handleSessionClick(sessionItem)}
-                >
-                  <Box sx={{ flex: 1, minWidth: 0 }}>
-                    <Typography variant="body1" fontWeight={500}>
-                      {sessionItem.groupName}
-                    </Typography>
-                    <Box sx={{ display: "flex", gap: 1, mt: 0.5, flexWrap: "wrap" }}>
-                      <Typography variant="caption" color="text.secondary">
-                        {formatTimestamp(sessionItem.timestamp)}
-                      </Typography>
-                      {sessionItem.studentCount !== null && sessionItem.studentCount !== undefined && (
-                        <>
-                          <Typography variant="caption" color="text.secondary">
-                            •
-                          </Typography>
-                          <Typography variant="caption" color="text.secondary">
-                            {sessionItem.studentCount} student{sessionItem.studentCount !== 1 ? "s" : ""}
-                          </Typography>
-                        </>
-                      )}
-                    </Box>
-                  </Box>
-                  <IconButton
-                    size="small"
-                    color="primary"
-                    onClick={(e) => {
-                      e.stopPropagation();
-                      handleSessionClick(sessionItem);
-                    }}
-                    sx={{ ml: 1 }}
-                  >
-                    <PlayArrowIcon />
-                  </IconButton>
-                </Box>
-              ))}
-            </Box>
-          </CardContent>
-        </Card>
-      )}
-
       {classrooms.length === 0 ? (
         <Card sx={{ textAlign: "center", py: 6 }}>
           <CardContent>
@@ -296,12 +208,7 @@ const ClassroomList = () => {
                 }}
               >
                 <CardContent sx={{ flexGrow: 1 }}>
-                  <Typography
-                    variant="h6"
-                    component="h2"
-                    gutterBottom
-                    sx={{ fontWeight: 600 }}
-                  >
+                  <Typography variant="h6" component="h2" gutterBottom sx={{ fontWeight: 600 }}>
                     {classroom.name}
                   </Typography>
                   <Box sx={{ mt: 2, display: "flex", flexDirection: "column", gap: 1 }}>
@@ -320,10 +227,7 @@ const ClassroomList = () => {
                   </Box>
                 </CardContent>
                 <CardActions sx={{ justifyContent: "space-between", px: 2, pb: 2 }}>
-                  <Button
-                    size="small"
-                    onClick={() => handleView(classroom._id)}
-                  >
+                  <Button size="small" onClick={() => handleView(classroom._id)}>
                     View
                   </Button>
                   <Box>
@@ -349,11 +253,14 @@ const ClassroomList = () => {
         </Grid>
       )}
 
+      <RecentConferences sessions={recentSessions} onSessionClick={handleSessionClick} />
+
       <Dialog open={deleteDialogOpen} onClose={handleDeleteCancel}>
         <DialogTitle>Delete Classroom</DialogTitle>
         <DialogContent>
           <Typography>
-            Are you sure you want to delete "{classroomToDelete?.name}"? This action cannot be undone.
+            Are you sure you want to delete "{classroomToDelete?.name}"? This action cannot be
+            undone.
           </Typography>
         </DialogContent>
         <DialogActions>
